+++ src/index.js
@@ -51,11 +51,6 @@ export function isEqual (value, other, compare) {
             return value !== 0 || 1 / value === 1 / other;
         }
 
-        // null 今早退出判断，以防进入 equalObject 判断
-        if (value === null || other === null) {
-            return false;
-        }
-
         const vType = type(value, true);
         const oType = type(other, true);
 
@@ -69,42 +64,24 @@ export function isEqual (value, other, compare) {
             return true;
         }
 
-        switch (vType) {
-        case 'Boolean':
-        case 'Date':
-        case 'Number':
+        // new Boolean|Number|Date
+        if (vType === 'Boolean' || vType === 'Number' || vType === 'date') {
             return +value === +other;
-        case 'String':
-        case 'regexp':
-            return '' + value === '' + other;
         }
 
-        if (vType !== 'array') {
-            // 存在 函数的情况
-            if (vType !== 'object' || oType !== 'object') {
-                return false;
-            }
-            
-            var vCtor = value.constructor;
-            var oCtor = other.constructor;
-            // value 和 other 构造函数都存在且不相等，那么他们就不该相等
-            if (
-                vCtor !== oCtor &&
-                !(
-                    type(vCtor, true) === 'function' &&
-                    type(oCtor, true) === 'function' &&
-                    'constructor' in value &&
-                    'constructor' in other
-                )
-            ) {
-                return false;
-            }
+        // new String | /123/ | new RegExp
+        if (vType === 'String' || vType === 'regexp') {
+            return String(value) === String(other);
         }
+
         if (vType === 'array') { // 数组判断
             return equalArray(value, other, compare);
-        } else { // 对象判断
+        }
+        if (vType === 'object') { // 对象判断
             return equalObject(value, other, compare);
         }
+
+        return value === other;
     };
 
     if(type(compare) === 'function') {
@@ -114,17 +91,5 @@ export function isEqual (value, other, compare) {
 }
 
 export function isEqualJSON(value, other, replacer = false) {
-    var vType = type(value);
-    var oType = type(other);
-    if (vType === 'regexp') {
-        value = {};
-    }
-    if (oType === 'regexp') {
-        other = {};
-    }
-    // -0 should not equal +0
-    if (value === 0 && 1 / value !== 1 / other) {
-        return false;
-    }
     return JSON.stringify(value, replacer) === JSON.stringify(other, replacer);
 }