+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # 变更日志
 
+## 0.2.0 / 2019-3-24
+
+- isEqual新增对new Date的支持
+- isEqual新增对new RegExp的支持
+- isEqual新增对regexp的支持
+
 ## 0.1.0 / 2019-3-20
 
 - 新增`isEqual()`